# Optimize snake case converter