Use more optimized algorithm for snake case conversion based on JSON.NET converter

https://github.com/JamesNK/Newtonsoft.Json/blob/master/Src/Newtonsoft.Json/Utilities/StringUtils.cs
See ToSeperatedCase method

Snake case converter can be hot path in applications that use it:
![image](https://user-images.githubusercontent.com/43544435/184496441-dc58e5a0-ae12-4c6c-8ea5-8db51e2ef521.png)

Current conversion method produce a lot of garbage and is not that fast

Benchmark repo
https://github.com/Kacper-Kondracki/SnakeCaseImprove

|                         Method |           Mean |       Error |      StdDev |   Allocated |
|------------------------------- |---------------:|------------:|------------:|------------:|
|             RemoraSingleString |     3,446.0 ns |     6.13 ns |     5.73 ns |     1,864 B |
|              Remora1000Strings | 4,788,162.1 ns | 5,290.28 ns | 4,948.54 ns | 1,861,700 B |
|            JsonNetSingleString |     1,227.3 ns |     5.29 ns |     4.95 ns |     1,240 B |
|             JsonNet1000Strings | 1,493,626.6 ns | 2,042.66 ns | 1,910.71 ns | 1,237,841 B |
| **JsonNetOptimizedSbSingleString** |     1,032.4 ns |     9.90 ns |     9.26 ns |       824 B |
|  **JsonNetOptimizedSb1000Strings** | 1,246,180.5 ns |   582.19 ns |   486.15 ns |   821,842 B |
|  JsonNetStackallocSingleString |       813.2 ns |     9.39 ns |     8.78 ns |       368 B |
|   JsonNetStackalloc1000Strings | 1,125,884.4 ns |   403.12 ns |   357.36 ns |   365,841 B |

Tested on random 128 sized alphanumerical strings

Proposed algorithm in this pr is **JsonNetOptimizedSb** - the same algorithm as JSON.NET, but preallocated StringBuilder to input size + (input size / 2). As you can see on benchmark, it saves a lot of memory.

**2.2x more efficient and around 3.8x faster** than current Remora implementation

It's more memory efficient, because not preallocated StringBuilder will always exceed input size and will cause to realloc to 2x of input size, compared to only 1.5x when preallocated in almost all cases.

Notice: In **very very very rare cases** converted string size can exceed 1.5x size, but **certainly** final string size will not exceed 2x of input size. It would be possible to implement correct stack based version, but  It's problematic to resize stack allocated span in safe C#.

In pr proposed version, StringBuilder will automatically reallocate to 3x size of input (because we preallocated 1.5x, and when it needs more space, it will allocate 2x of its current size), but this happens only with very odd inputs. In practical scenarios it will probably never happen.

Additional benchmark - input = "ThisIsSomeField"
|                         Method |     Mean |   Error |  StdDev | Allocated |
|------------------------------- |---------:|--------:|--------:|----------:|
|             RemoraSingleString | 342.2 ns | 2.88 ns | 2.56 ns |     400 B |
|            JsonNetSingleString | 140.2 ns | 0.63 ns | 0.56 ns |     272 B |
| JsonNetOptimizedSbSingleString | 116.3 ns | 0.62 ns | 0.55 ns |     184 B |