@AraHaan 2022-08-13
I think I like the stackalloc solution much more, the least amount of memory used while not sacrificing performance, the better for all users of the Remora.Rest (even Remora.Discord could benefit from it).