@Kacper-Kondracki 2022-08-14
@AraHaan The stackalloc version allocates fixed buffer of 1.5 of input size, but for rare odd inputs, it throws out of bound exception.

I'm thinking about solution that saves memory the most, conversion in two passes, first pass calculates the final buffer size, second actually converts the string. That way we will save the most memory

Estimated performance penalty in my opinion = 2x, but I think it will still be faster than current remora implementation.
I will prepare benchmarks later.

Honestly after thinking twice, the memory usage reduction will be better than performance.

@Nihlus What is your opinion on this?