@Nihlus 2022-08-17
This is a great change, and an optimization pass is sorely needed in lots of places of the library.

As this triggered an impromptu optimization contest on the server, I'd like to see @VelvetThePanda's version implemented in this PR (or whichever version turns out to be faster than that one!).