@Nihlus 2022-08-18
I went ahead and implemented a version roughly based on JSON.NET's stackalloc variant, modified with my own optimizations and wrapped in some safety features and additional flexibility. The final version just barely differs in execution time and exactly matches it in allocation size.

```
|             Method |         Value |      Mean |    Error |   StdDev | Ratio | RatioSD | Allocated |
|------------------- |-------------- |----------:|---------:|---------:|------:|--------:|----------:|
|             Remora | OnceUponATime | 200.72 ns | 4.010 ns | 3.939 ns |  1.00 |    0.00 |     288 B |
|            JsonNet | OnceUponATime |  67.27 ns | 1.388 ns | 2.927 ns |  0.34 |    0.02 |     160 B |
| JsonNetOptimizedSb | OnceUponATime |  72.28 ns | 0.830 ns | 0.736 ns |  0.36 |    0.01 |     168 B |
|  JsonNetStackalloc | OnceUponATime |  51.04 ns | 1.058 ns | 1.550 ns |  0.26 |    0.01 |      56 B |
|      JaxStackalloc | OnceUponATime |  53.06 ns | 0.497 ns | 0.440 ns |  0.26 |    0.01 |      56 B |
```

Thanks for the PR and the work it encouraged! I'll close this as completed.